refactored input output data files to correct location
