+++ src/main/java/OutputData/GCalEventOutputData.java
@@ -1,4 +1,4 @@
-package use_case;
+package OutputData;
 
 import com.google.api.client.util.DateTime;
 import com.google.api.services.calendar.Calendar;
