+++ src/main/java/use_case/GCalEventOutputBoundary.java
@@ -1,5 +1,7 @@
 package use_case;
 
+import OutputData.GCalEventOutputData;
+
 public interface GCalEventOutputBoundary {
     void prepareSuccessView(GCalEventOutputData gCalEventOutputData);
     void prepareFailView(String error);
